streamlit/requirements.txt
streamlit==1.37.0
pandas==2.2.2
psycopg2-binary==2.9.9
streamlit-autorefresh==1.0.1
plotly==6.5.2